sadf: SVG: Don't insert id tag if activity not displayed
When using both "showtoc" and "skipempty" options, e.g.:

$ sadf -g -O showtoc,skipempty -- -A > output.svg

clicking on the link pointing at an activity with no graphs displayed
(because all its statistics were zero) would move to the first
following graph... which is not the graph corresponding to the link!
This was because even if the graphs don't get displayed, sadf inserted
the id tag in the SVG file, e.g.:

<g id="g34-0" transform="translate(0,7870)">
</g>
<g id="g8-0" transform="translate(0,7870)">
<rect x="0" y="0" height="300" width="1050"/>

In the example above, clicking on the link to display graph whose id is
"g34-0" would display graph with id "g8-0" instead.

This patch fixes the problem.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>