sadf: SVG: Fix possible negative value in graph coordinate
dt interval may sometimes be greater by one unit than the timetag
interval because of a previous rounding operation.

When this happens, the coordinate gets a crazy value which is obviously
wrong, e.g.:

<rect x="18446744073709551615" y="0.00"...

So make a comparison between dt and timetag values, and use 0 if dt
happens to be greater than timetag.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>