sar: Better detect if a disk has been unregistered then registered again
If a disk was mounted read-only then sar might not detect that a disk
had been unregistered then registered again. This is because detection
is based on counters (like number of written sectors) going down. Yet
with a read-only device, the number of written sectors is always 0 and
is never decremented.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>