sadf: Time used by options -s/-e no longer depends on the timezone value
The time specified with options -s and -e used by sadf has no longer to
be expressed in local time. This time is now consistent with that
displayed by sadf, e.g.: If sadf displays its timestamps in local time
then the time specified with options -s/-e is considered as expressed in
local time. If sadf displays its timestamps in UTC then the time used
with options -s/-e is also considered as expressed in UTC.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>