sar/sadc: Add stable identifier support for disks statistics
This patch adds new fields to stats_disk structure to save a stable
identifier for each block device (see issue #195).
A stable identifier is a name that should not change across reboots for
the same physical device.
At the present time this stable identifier is the WWN (World Wide Name)
id that is read from /dev/disk/by-id if it exists for the device.
If it doesn't exist then we fall back on using the pretty name (sda,
sda1, etc.).
The stable identifier is always collected by sadc when disks statistics
are collected (sadc option "-S DISK | XDISK").
It can be printed by sar (or sadf) with the option "-j SID" (SID stands
for Stable IDentifier).

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>