sysstat-12.1.7
sysstat version 12.1.7 final packaging.
lsm and spec files updated.
Changelog added.

The main new feature brought by this version is the support for stable
identifiers for sar disks statistics. This stable identifier (which
won't change across reboots) is the WWN (World-Wide Name) of the device.
The sar and sadf commands included in this version are also ready to read
new formats of binary data files which could arrive in a (hopefully)
distant future.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>