sar/sadf: Make sure we can read file headers structures from older versions
Reading the file_header or file_activity structures from a daily data
file from an older version where fewer fields were present might be a
problem as we allocate buffers based on the size of the structure from
the file. This buffer would ithen be to small for the remap_struct()
function to remap the fields to current (larger) structures.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>