pidstat: Remove extra space at end of command name
When pidstat was used with option "-l" to display the command name and
all its arguments, a space was added at the end after all the arguments.
Remove it as it is unneeded and it may prevent the user from displaying
lines based on a pattern located at the end of line
(e.g. pidstat -l -C "pattern$" ...)

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>