pidstat: Major code refactoring
Now use a linked list and dynamic allocatin to save tasks being
monitored.
Also fix several bugs that were lurking in previous pidstat version,
like threads that were displayed as attached to the wrong thread leader
if their real leader was not displayed because it was not an active
task.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>