sa2: Wait for a random delay before running
Add a new option ("delay_range=") to configure script to tell sa2 script
to wait for a random delay in the indicated range before running.
This delay (expressed in seconds) is aimed at preventing a massive I/O
burst at the same time on VM sharing the same storage area.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>