cifsiostat,iostat,sar: Add --pretty option
This option makes the report easier to read by a human.
Please note that for sar, option -p is now equivalent to this one.
Use this option to display device names or network interface names
on the right of the report instead of the left (which can be
particularly useful for long names).

Also option -h is now equivalent to "--pretty --human".

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>