sar: Don't use IFNAMSIZ value from <net/if.h>
IFNAMSIZ value is used to set the maximum length of the network
interface name, and for the definition of structures used to save
data in the binary data files (saDD).
We cannot allow that a new kernel value for IFNAMSIZ changes the format
of saDD binary data files. So now use a constant value of 16 (which is
the present value of IFNAMSIZ).

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>