Update sadf to use metadata matching the same metrics from PCP
When operating with the -l option in sadf we want to generate PCP
archives that match the PCP archives generated by pmlogger(1) as
closely as possible.  The advantage to doing this is that the live
and historical modes to the PCP tools become interchangeable; i.e.
pmchart views, pmie rules, grafana-pcp dashboards, and so on, all
expect metrics named according to the PCP conventions.

The reason sadf is not already doing this (AIUI), is that it was
originally based on the sar2pcp(1) utility.  This tool (which we
plan to deprecate in PCP in favour of the sysstat functionality)
used the rate-converted counter values as its starting point and
not the raw counters we have access to within sysstat itself.

I've audited all of the metrics being added to PCP archives with
the sadf -l option now, and all now align with current metadata
in PCP.  I've also added a few uname(2) metrics from the record
header (in the way hinv.ncpu was already being used).