mpstat: Don't use read_stat_irq() function
mpstat used read_stat_irq() function from rd_stats.c to get the total
number of interrupts among all processors.
This function is to be heavily modified to deal with sar's new
interrupts statistics report. So create a new function which will be
used only by mpstat and include it in its code.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>