A_IRQ: Change syntax for option -I
Option -I was followed by a list of interrupts or the SUM/ALL keywords.
Now option -I doesn't need to be followed par additional parameters.
Keywords SUM and ALL may still be used (but the ALL keyword will be
ignored).
The list of interrupts will be entered using a new option: --int=

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>