mpstat : Add reset parameter to salloc_mp_struct()
To reset stats structs memory, add parameter to salloc_mp_struct().

This will be used in the next commit for introducing -H option for
handling vCPU physical hotplugs.

Use realloc calls which will help in resizing the structs also.
if reset is TRUE, memory will be cleared else not cleared

Signed-off-by: Shaleen Bathla <shaleen.bathla@oracle.com>