Try to avoid displaying healthy metrics values in "red" (#177)
Percentage values are now displayed in specific colors when they are
greater or lower than a predefined limit depending on the metric's
meaning.
This means that, e.g. CPU %idle value will no longer be printed in red
if the CPU is actually idle.

Specific colors are now used when:
* values are greater than 90% (or lower than 10%),
* values are in the range 75% to 90% (or 25% to 10%).

This is probably still not perfect but should eliminate many questions.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>