sar/sadf: Rework code related to options -s/-e
Timestamps entered as a number of seconds since the Epoch are now used
as they are and no longer converted to a broken time representation, so
that even month and year are taken into account when current record time
is compared to those timestamps.

The record which is exactly at the ending time is now displayed.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>