sar: Reallocate min/max buffers when necessary
The xdev_list contains all the devices (items) which have been found
until now, even those which have been now unregistered. As a
consequence, this list may grow larger than the number of slots
allocated in spmin and spmax buffers used to save min and max values
(the number of slots allocated by default is the same as the maximum
number of different devices found in a single sample, which may be
smaller than the number of all the devices found until now).
So reallocate the spmin and spmax buffers when they are too small.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>