sar/sadf: [A_NET_ETCP]: Rename retrans/s field to retrseg/s
A_NET_ETCP report (displayed by "sar -n ETCP") has a column named
retrans/s. This name is also used in the A_NET_NFS report, which can be
confusing. So rename it to retrseg/s.
This field is also renamed in the output of sadf (PPC, DB and SVG output
formats).

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>