configure.ac: fix erroneous bashisms
An autoconf generated script is designed to work with POSIX sh, and
contains a /bin/sh shebang. As a result, it *cannot* assume it will be
run with bash, as it won't be.

POSIX sh provides no way to extend the contents of a variable. A
bash-specific extension is to use the += notation.

Instead, the standard (albeit somewhat more wordy) POSIX compatible
approach is to expand the previous contents into a regular assignment.
Do so.

Fixes: https://bugs.gentoo.org/923989

 * QA Notice: Abnormal configure code
 *
 * ./configure: 4927: CFLAGS+= -m32: not found